requirements_pip.txt
flask
flask-cors
python-dotenv
pandas
pillow
tensorflow
numpy
scikit-learn
xgboost
opencv-python
matplotlib
seaborn
